pyproject.toml
[project]
name = "poly-glide-template"
version = "0.0.1"
description = "This is a template project to demonstrate how to integrate with PolyAPI in a git-centric python workflow."
authors = [{ name = "My Name", email = "" }]
dependencies = [
    "polyapi-python==0.3.1.dev13",
    "pre-commit==4.0.1",
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

[project.urls]
Homepage = ""
